+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+
+- optimization, prevent timeline to fetch multiple times the API
 - Internal email delivery is more reliable, with better error handling, reporting and logging.
 
 ### Fixed